// src/types/Tournament-Ranking/TournamentRanking.ts
export interface TournamentRanking {
    idPlayer: number;
    name: string;
    lastname: string;
    position: number;
    points: number;
    photo: string;
    wonMatches: number;
    lostMatches: number;
    playedMatches: number;
    setsWon: number;
    setsLost: number;
    setsDifference: number;
}